third_party/blink/web_tests/platform/generic/crypto/subtle/pbkdf2/deriveKey-aes-expected.txt

Tests encrypt/decrypt for AES-CBC

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "PBKDF2"
PASS tmpKey.usages.join(',') is "deriveKey,deriveBits"
PASS dkey.type is "secret"
PASS typeof dkey.extractable is 'boolean'
PASS dkey.extractable is true
PASS dkey.algorithm.name is "AES-CBC"
PASS dkey.algorithm.length is 128
PASS dkey.usages.join(',') is "encrypt"
PASS: Derivedkey should be [0c60c80f961f0e71f3a9b524af601206] and was
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "PBKDF2"
PASS tmpKey.usages.join(',') is "deriveKey,deriveBits"
PASS dkey.type is "secret"
PASS typeof dkey.extractable is 'boolean'
PASS dkey.extractable is true
PASS dkey.algorithm.name is "AES-CBC"
PASS dkey.algorithm.length is 128
PASS dkey.usages.join(',') is "encrypt"
PASS: Derivedkey should be [ea6c014dc72d6f8ccd1ed92ace1d41f0] and was
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "PBKDF2"
PASS tmpKey.usages.join(',') is "deriveKey,deriveBits"
PASS dkey.type is "secret"
PASS typeof dkey.extractable is 'boolean'
PASS dkey.extractable is true
PASS dkey.algorithm.name is "AES-CBC"
PASS dkey.algorithm.length is 128
PASS dkey.usages.join(',') is "encrypt"
PASS: Derivedkey should be [4b007901b765489abead49d926f721d0] and was
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "PBKDF2"
PASS tmpKey.usages.join(',') is "deriveKey,deriveBits"
PASS dkey.type is "secret"
PASS typeof dkey.extractable is 'boolean'
PASS dkey.extractable is true
PASS dkey.algorithm.name is "AES-CBC"
PASS dkey.algorithm.length is 128
PASS dkey.usages.join(',') is "encrypt"
PASS: Derivedkey should be [3d2eec4fe41c849b80c8d83662c0e44a] and was
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "PBKDF2"
PASS tmpKey.usages.join(',') is "deriveKey,deriveBits"
PASS dkey.type is "secret"
PASS typeof dkey.extractable is 'boolean'
PASS dkey.extractable is true
PASS dkey.algorithm.name is "AES-CBC"
PASS dkey.algorithm.length is 128
PASS dkey.usages.join(',') is "encrypt"
PASS: Derivedkey should be [56fa6aa75548099dcc37d7f03425e0c3] and was
PASS successfullyParsed is true

TEST COMPLETE
